fix сode issues
