+++ packages/client/src/store/auth/auth.slice.ts
@@ -33,7 +33,6 @@ export const authApi = createApi({
         method: 'POST',
         body: data,
         responseHandler: response => {
-          console.log('2', response );
           const isJson = response.headers
             .get('Content-Type')
             ?.includes('application/json');
